# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tracking-System
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/password/forgot:
    post:
      summary: 'Forgot Password'
      operationId: forgotPassword
      description: 'Generate a password reset token for the given email address.'
      parameters: []
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: schaden.isaias@example.com
              required:
                - email
      security: []
  /api/password/check:
    post:
      summary: 'Check Reset Token'
      operationId: checkResetToken
      description: 'Check if the given password reset token is valid.'
      parameters: []
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: ''
                  example: earum
              required:
                - token
      security: []
  /api/password/reset:
    post:
      summary: 'Reset Password'
      operationId: resetPassword
      description: "Reset the user's password using the given password reset token."
      parameters: []
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: ''
                  example: sunt
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: adrien.kihn@example.org
                password:
                  type: string
                  description: ''
                  example: velit
              required:
                - token
                - email
                - password
      security: []
...
  /api/user/change-role:
    post:
      summary: 'Change Role'
      operationId: changeRole
      description: 'Update user role'
      parameters: []
      responses: {  }
      tags:
        - 'User Controller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: ''
                  example: quia
                role:
                  type: string
                  description: ''
                  example: Assigner
                  enum:
                    - Assignee
                    - Assigner
              required:
                - user_id
                - role
      security: []
...
  /api/user/password/change:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: "Update the authenticated user's password."
      parameters: []
      responses: {  }
      tags:
        - 'User Controller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: drake16@example.net
                old_password:
                  type: string
                  description: ''
                  example: sit
                new_password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: hzculgxapbehcbpvqqmjgerxeicpodibfbpcydevgvzvmqyqgdzrmjlekvioyasmgrektxmwom
              required:
                - email
                - old_password
                - new_password
      security: []
...
  /api/department/assignee:
    post:
      summary: 'Show Assignee'
      operationId: showAssignee
      description: 'Returns a list of users who have the "assignee" role in the specified department'
      parameters: []
      responses: {  }
      tags:
        - 'User Controller'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                department_id:
                  type: string
                  description: ''
                  example: null
      security: []
  /api/department/assigner:
    post:
      summary: 'Show Assigner'
      operationId: showAssigner
      description: 'Returns a list of users who have the "assigner" role in the specified department'
      parameters: []
      responses: {  }
      tags:
        - 'User Controller'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                department_id:
                  type: string
                  description: ''
                  example: null
      security: []
...